requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
SQLAlchemy==1.4.46  # Using an older version for better compatibility
Werkzeug==2.2.3
Jinja2==3.1.2
WTForms==3.0.1
email-validator==1.3.1
python-dotenv==1.0.0
pytest==7.3.1
pytest-flask==1.2.0
coverage==7.2.2
flake8==6.0.0
black==23.3.0
openai==1.97.1